requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
alembic==1.7.7
    # via -r requirements.in
anyio==3.5.0
    # via starlette
asgiref==3.5.1
    # via uvicorn
asyncpg==0.25.0
    # via -r requirements.in
black==22.3.0
    # via -r requirements.in
click==8.1.3
    # via
    #   black
    #   uvicorn
ecdsa==0.17.0
    # via python-jose
fastapi==0.75.2
    # via -r requirements.in
flake8==4.0.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
importlib-metadata==4.11.3
    # via alembic
importlib-resources==5.7.1
    # via alembic
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via mako
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
passlib==1.7.4
    # via -r requirements.in
pathspec==0.9.0
    # via black
platformdirs==2.5.2
    # via black
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycodestyle==2.8.0
    # via flake8
pydantic==1.9.0
    # via fastapi
pyflakes==2.4.0
    # via flake8
python-dotenv==0.20.0
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.36
    # via
    #   -r requirements.in
    #   alembic
starlette==0.17.1
    # via fastapi
tomli==2.0.1
    # via black
typing-extensions==4.2.0
    # via
    #   black
    #   pydantic
uvicorn==0.17.6
    # via -r requirements.in
zipp==3.8.0
    # via
    #   importlib-metadata
    #   importlib-resources
